downloader/requirements.txt
yt-dlp>=2024.1.7
redis==5.0.1
aiosqlite==0.19.0
python-dotenv==1.0.0
python-telegram-bot==20.7
Pillow==10.1.0
